Added certificate search specification

+++ module4/gift-certificate-system/api/src/main/java/com/epam/esm/constants/AppConstants.java
@@ -1,7 +1,7 @@
 package com.epam.esm.constants;
 
 public class AppConstants {
-    public static final Long DEFAULT_PAGE_NUMBER = 1L;
-    public static final Long DEFAULT_PAGE_SIZE = 20L;
+    public static final int DEFAULT_PAGE_NUMBER = 1;
+    public static final int DEFAULT_PAGE_SIZE = 20;
 
 }
