+++ module4/gift-certificate-system/api/src/main/java/com/epam/esm/service/OrderService.java
@@ -6,10 +6,10 @@ import com.epam.esm.model.Order;
 import java.util.List;
 
 public interface OrderService {
-    List<OrderDTO> getAllOrders(Long page, Long pageSize);
+    List<OrderDTO> getAllOrders(int page, int pageSize);
     OrderDTO getOrderById(Long id);
     OrderDTO makeOrder(Long userId, Long certificateId);
     void removeOrder(Long orderId);
-    List<OrderDTO> getUserOrders(Long userId, Long page, Long pageSize);
+    List<OrderDTO> getUserOrders(Long userId, int page, int pageSize);
 
 }
